+++ Cargo.toml
@@ -25,7 +25,7 @@ mmap = ["fs", "memmap", "task_unstable"]
 log = "0.4.0"
 cfg-if = "1.0.0"
 futures-lite = { version = "1.7.0", optional = true }
-futures-timer = { version = "2.0.2", optional = true }
+futures-timer = { version = "3.0.0", optional = true }
 async-std = { version = "1.6.0", default-features = false, optional = true }
 async-io = { version = "1.1.2", optional = true }
 async-fs = { version = "1.3.0", optional = true }
@@ -53,8 +53,8 @@ bytes = "0.5.6"
 lazy_static = "1.2.0"
 num_cpus = "1.10.1"
 futures-util = { version = "0.3.5", features = ["sink"] }
-async-lock = "1.1.0"
+async-lock = "2.0.0"
 tokio-util = { version = "0.3.1", features = ["codec", "compat"] }
 flv-util = { version = "0.5.0", features = ["fixture"] }
 fluvio-test-derive = { path = "async-test-derive", version = "0.1.0" }
-fluvio-future = { path = ".", features = ["net", "fixture", "timer", "fs","tls","native2_tls"] }
+fluvio-future = { path = ".", features = ["net", "fixture", "timer", "fs"] }
